requirements.txt
alembic==1.10.2
aniso8601==9.0.1
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.2
fernet==1.0.1
Flask==2.2.3
Flask-JWT-Extended==4.5.3
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
packaging==23.0
pyaes==1.6.1
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.0.2
pytz==2022.7.1
six==1.16.0
SQLAlchemy==2.0.6
typing_extensions==4.5.0
Werkzeug==2.2.3
